Refactors reliability factors into metrics module
