+++ src/easydiffraction/analysis/minimization.py
@@ -13,7 +13,7 @@ from easydiffraction.core.objects import Parameter
 from easydiffraction.experiments.experiments import Experiments
 from easydiffraction.sample_models.sample_models import SampleModels
 
-from ..analysis.reliability_factors import get_reliability_inputs
+from easydiffraction.analysis.fitting.metrics import get_reliability_inputs
 from .fitting.results import FitResults
 from .minimizers.minimizer_factory import MinimizerFactory
