+++ tests/unit/analysis/test_reliability_factors.py
@@ -2,12 +2,12 @@ from unittest.mock import Mock
 
 import numpy as np
 
-from easydiffraction.analysis.reliability_factors import calculate_r_factor
-from easydiffraction.analysis.reliability_factors import calculate_r_factor_squared
-from easydiffraction.analysis.reliability_factors import calculate_rb_factor
-from easydiffraction.analysis.reliability_factors import calculate_reduced_chi_square
-from easydiffraction.analysis.reliability_factors import calculate_weighted_r_factor
-from easydiffraction.analysis.reliability_factors import get_reliability_inputs
+from easydiffraction.analysis.fitting.metrics import calculate_r_factor
+from easydiffraction.analysis.fitting.metrics import calculate_r_factor_squared
+from easydiffraction.analysis.fitting.metrics import calculate_rb_factor
+from easydiffraction.analysis.fitting.metrics import calculate_reduced_chi_square
+from easydiffraction.analysis.fitting.metrics import calculate_weighted_r_factor
+from easydiffraction.analysis.fitting.metrics import get_reliability_inputs
 
 
 def test_calculate_r_factor():
